remove author's ad to conform with the new NPM policy
While I don't really agree with the new NPM policy as reported here: https://www.zdnet.com/article/npm-bans-terminal-ads/
I do think it could give grounds for removal/ban of this package. If this were to happen, we'd get another left-pad debacle.
So maybe let's remove this just to be on the safe side with NPM shall we?

PS: If you have a tweet where you announce looking for a job or portfolio I'd be happy to share link to that on my social media.