Remove copyright from executable code
This winds up in the minified distribution everywhere this library is used instead of the usual inclusion on a page with a list of the open source software used (i.e. licenses page). It's a minor thing, but multiplied by millions and millions of instances of the notice across likely billions of requests.