limit postinstall ad to show only one time
When installing multiple packages that depend on core-js, the postinstall message will pop up mutiple times which will add unnecessary verbosity to the terminal log. I know #767, #548, #797, #757, #781 ,#729, #708, refer to this problem and since @zloirock refuses to remove this, lets limit it to one time so that verbosity gets truncated to the least.