Cleanup postinstall message
There is no longer a need for funding this project, or hiring the author, seeing as he is in jail. The project has already verifiably received $5,000 USD to help with operations, so it is unclear where this money is going or who it is helping if people choose to donate.