use a smaller value for "length" when testing array "ToLength"
Using the value -1 for the length of the object in the
`arrayMethodUsesToLength` tests causes implementations to test against a
very large "array". Use a smaller value to not cause those
implementations to hang. This change allows all the uses of
`arrayMethodUsesToLength` to use the actual function they are testing
rather than a "proxy". In addition, change the testing function to use
`apply` rather than `call` so that implementations that count the number
of arguments will use the default values for any arguments that are
optional rather than trying to interpret `undefined` incorrectly.